+++ projects/io/src/app/core/pages/forms/forms.module.ts
@@ -1,16 +1,15 @@
 
 import { NgModule } from "@angular/core";
 import { Routes, RouterModule } from "@angular/router";
-import { FileInputPage } from "./file-input.page";
+import { FormsPage } from "./forms.page";
 import { FlexLayoutModule } from "@angular/flex-layout";
 import { CommonModule } from "@angular/common";
-import { NgEzCodePrettifyModule, NgEzFileModule } from '@ngez/core';
+import { NgEzCodePrettifyModule } from '@ngez/core';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { ReactiveFormsModule } from "@angular/forms";
 
 const routes : Routes = [{
     path: '',
-    component: FileInputPage
+    component: FormsPage
 }];
 
 @NgModule({
@@ -19,9 +18,7 @@ const routes : Routes = [{
         MatProgressSpinnerModule,
         FlexLayoutModule,
         RouterModule.forChild(routes),
-        CommonModule,
-        NgEzFileModule,
-        ReactiveFormsModule],
-    declarations: [FileInputPage]
+        CommonModule],
+    declarations: [FormsPage]
 })
-export class FileInputModule{}
+export class FormsModule{}